[refactor] #4 extract price info section into separate view
